refactor: split manage_routes into focused submodules
Rebased onto main and split the ~2400-line manage_routes.py into a
package covering config, draft, LLM, knowledge, reindex helpers, and
runtime apply logic. Preserves backward-compatible imports including
draft locking, secret redaction, and deferred reindex helpers added on
main.
