+++ tests/unit/test_manage_publish_sync.py
@@ -46,11 +46,12 @@ def test_publish_syncs_draft_with_published_knowledge_flags(monkeypatch):
         params={"agent_id": "test-agent"},
         json={
             "config": {
+                "agent": {"name": "Test Agent"},
                 "knowledge": {
                     "enabled": True,
                     "agent_level_enabled": True,
                     "session_level_enabled": False,
-                }
+                },
             }
         },
     )
@@ -87,11 +88,12 @@ def test_publish_syncs_draft_with_published_agent_level_disabled(monkeypatch):
         params={"agent_id": "test-agent"},
         json={
             "config": {
+                "agent": {"name": "Test Agent"},
                 "knowledge": {
                     "enabled": True,
                     "agent_level_enabled": False,
                     "session_level_enabled": True,
-                }
+                },
             }
         },
     )
